Slight edit.
So basically I just switched where x starts at x=h and ends at nh where
I defined h=1/(n+1) this allows the end points to be not included in
the x vector when it is calculated. Thus the f should match up to the
proper values making it solve this shit.